+++ src/Dotnet.Samples.AspNetCore.WebApi/Program.cs
@@ -122,3 +122,8 @@ if (app.Environment.IsDevelopment())
 app.MapControllers();
 
 await app.RunAsync();
+
+public partial class Program
+{
+    protected Program() { }
+}